requirements.txt
anyio==3.6.1
asgiref==3.5.2
attrs==21.4.0
Automat==20.2.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
constantly==15.1.0
cryptography==37.0.2
cssselect==1.1.0
decorator==5.1.1
fastapi==0.78.0
filelock==3.7.0
fire==0.3.1
Flask==2.1.2
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
geojson==2.5.0
geojson-pydantic==0.3.4
geopy==2.2.0
h11==0.13.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.4
incremental==21.3.0
itemadapter==0.6.0
itemloaders==1.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
kanjize==1.0.0
lxml==4.8.0
MarkupSafe==2.1.1
motor==3.0.0
mysqlclient==2.1.0
normalize-japanese-addresses==0.0.5
numpy==1.22.4
pandas==1.4.2
parsel==1.6.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
PyDispatcher==2.0.5
pymongo==4.1.1
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
queuelib==1.6.2
ratelim==0.1.6
requests==2.27.1
requests-file==1.5.1
scipy==1.8.1
Scrapy==2.6.1
scrapy-mongodb==0.12.0
service-identity==21.1.0
simple-geocoding==0.0.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.19.1
swagger-ui-py==21.12.8
termcolor==1.1.0
tldextract==3.3.0
Twisted==22.4.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
w3lib==1.22.0
Werkzeug==2.1.2
zipp==3.8.0
zope.interface==5.4.0
